refactor(core): extract intent models from intent_analyzer.py to _intent_models.py
Extrai os 5 símbolos de modelo (`IntentType`, `IntentCategory`,
`IntentPattern`, `IntentAnalysisResult`, `IntentCacheEntry`) do módulo
`deile/core/intent_analyzer.py` (853 LOC) para um novo módulo-folha
`deile/core/_intent_models.py` (stdlib-only, sem dependências externas).

O módulo original mantém `IntentAnalyzer` e o singleton
`get_intent_analyzer`, além de um shim de re-export (`__all__`) e um
import `from ._intent_models import ... # noqa: F401` que preserva
compatibilidade retroativa: todos os importadores existentes
(`intent_tier_mapper.py`, `agent.py`, testes) continuam funcionando sem
alteração.

ACs verificados:
- `wc -l _intent_models.py` = 135 (≤ 135)
- `wc -l intent_analyzer.py` = 735 (≤ 735)
- 22 testes impactados: 22 passed
